GG-34170 More refactoring
