+++ modules/platforms/cpp/network/CMakeLists.txt
@@ -31,6 +31,7 @@ set(SOURCES
         src/network/network.cpp
         src/network/ssl/secure_data_filter.cpp
         src/network/ssl/secure_socket_client.cpp
+        src/network/ssl/secure_utils.cpp
         src/network/ssl/ssl_gateway.cpp)
 
 if (WIN32)
