+++ modules/platforms/cpp/network/include/ignite/network/network.h
@@ -22,6 +22,7 @@
 #include <ignite/common/common.h>
 #include <ignite/network/socket_client.h>
 #include <ignite/network/async_client_pool.h>
+#include <ignite/network/ssl/secure_configuration.h>
 
 namespace ignite
 {
@@ -42,14 +43,11 @@ namespace ignite
             /**
              * Make secure socket for SSL/TLS connection.
              *
-             * @param certPath Certificate file path.
-             * @param keyPath Private key file path.
-             * @param caPath Certificate authority file path.
+             * @param cfg Configuration.
              *
              * @throw IgniteError if it is not possible to load SSL library.
              */
-            IGNITE_IMPORT_EXPORT SocketClient* MakeSecureSocketClient(const std::string& certPath,
-                const std::string& keyPath, const std::string& caPath);
+            IGNITE_IMPORT_EXPORT SocketClient* MakeSecureSocketClient(const SecureConfiguration& cfg);
         }
 
         /**
