+++ modules/platforms/cpp/network/project/vs/network.vcxproj
@@ -208,6 +208,7 @@
     <ClInclude Include="..\..\os\win\src\network\win_async_worker_thread.h" />
     <ClInclude Include="..\..\src\network\ssl\secure_socket_client.h" />
     <ClInclude Include="..\..\src\network\ssl\ssl_gateway.h" />
+    <ClInclude Include="..\..\src\network\ssl\secure_utils.h" />
     <ClInclude Include="..\..\src\network\async_client_pool_adapter.h" />
     <ClInclude Include="..\..\src\network\error_handling_filter.h" />
     <ClInclude Include="..\..\src\network\tcp_socket_client.h" />
@@ -230,6 +231,7 @@
     <ClCompile Include="..\..\src\network\ssl\secure_data_filter.cpp" />
     <ClCompile Include="..\..\src\network\ssl\secure_socket_client.cpp" />
     <ClCompile Include="..\..\src\network\ssl\ssl_gateway.cpp" />
+    <ClCompile Include="..\..\src\network\ssl\ssl_utils.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
